WebGL: Big change. ShaderResourceLayout is no longer required during shader creation
Instead, the work of binding uniform locations to uniform buffer offsets is done by the render pipeline. This now means that the resource layout does not need to be known at shader creation time, just at pipeline creation time.

So the same shader could be used for instanced and non-instanced rendering. This also makes it easier to create shaders during resource loading.

Material now takes a shader AND a resource layout, which makes sense.