Use ThreeJS' Object3D
This should make it easier to integrate ThreeJS animation system, at the cost of extra bundle size (20KB!) and lower performance. It uses a separate math library which seems very inferior (performance and flexibility wise) to gl-matrix. The vector/quat types are not POD, so it's harder to serialize-deserialize to files or GPU buffers. I now have to convert between the vector types in several places.

Hopefully I can reduce some of the extra code that is pulled in with Object3D in the future by exluding unneeded functions, but this may not be possible.