Fixed bad bug where AckInfo objects were being shared, causing invalid RTT times
This was causing the client to only sync to the server clock time once. I originally meant for the AckInfo object sent to the event to be reused, and the receiver should copy it. In hindsight that's definitely not worth the mental overhead, and mistakes like this were bound to happen.