Greatly improve the accuracy of serverTime computation
The previous guess would be off by an average of simDt/2 = 8ms. The server time was being set at a random point between two render frames (and two clock ticks). At the instant that syncToServerTime() was called, the guess was correct. But on the next clock tick, serverTime = serverTime + dt. Where dt is the delta between the two clock ticks, NOT the delta between the current clock tick and the time that syncToServerTime() was called. If sync was called just before the next clock tick, it would be off by an entire dt (16ms). Now we do a clock tick during sync, so the next dt will be correct.