NetClient computes extremely accurate serverTime (within 0.05 - 1.5ms)
The issue was that we were using the server frame to compute the server time at which we received the packet, but that can be off by as much as 15.99ms, since fixed frames are "produced" by the server time. Now we encode the current server time in the frame and use the new accurate ping measurement to get the one way trip time. Together this produces an extremely accurate server time measurement.