Created separate avatarExport and avatarRig.blend files
Export contains only the necessary and cleaned versions of the animations that the game will use, as well as the game's version of the armature.

The Rig file contains the full rig (with IK controls) and all animations.

New animations can be "appended" to the Export file for inclusion in the game.