AvatarController handles roll attack separately, rolls in a smooth curve
The curve is defined by a cubic hermite spline (start/end and two tangents), and can be tweaked by modifying the tangents. The simulation saves the starting point of the attack, and then re-evaluates the curve each frame based on the new target position.

Currently debug rendering of the curve is still enabled for easier tweaking.